+++ lmcache/cli/commands/bench/__init__.py
@@ -26,6 +26,7 @@ from lmcache.cli.commands.bench.engine_bench.stats import (
     StatsCollector,
 )
 from lmcache.cli.commands.bench.engine_bench.workloads import create_workload
+from lmcache.cli.commands.test_cache import TestCacheCommand
 from lmcache.logging import init_logger
 
 logger = init_logger(__name__)
@@ -52,9 +53,10 @@ class BenchCommand(BaseCommand):
         inner = parser.add_subparsers(
             dest="bench_target",
             required=True,
-            metavar="{engine}",
+            metavar="{engine,kvcache}",
         )
         self._register_engine(inner)
+        self._register_kvcache(inner)
 
     def _register_engine(
         self,
@@ -312,8 +314,42 @@ class BenchCommand(BaseCommand):
 
         parser.set_defaults(func=self.execute)
 
+    # ------------------------------------------------------------------
+    # kvcache bench target — end-to-end MP cache sanity test
+    # ------------------------------------------------------------------
+
+    def _register_kvcache(
+        self,
+        subparsers: argparse._SubParsersAction,
+    ) -> None:
+        """Register ``lmcache bench kvcache`` subcommand.
+
+        Arguments and execution are delegated to
+        :class:`~lmcache.cli.commands.test_cache.TestCacheCommand` so
+        the per-chunk store/retrieve/checksum logic lives in a single
+        place.
+        """
+        parser = subparsers.add_parser(
+            "kvcache",
+            help=TestCacheCommand().help(),
+            description=(
+                "End-to-end sanity test for the LMCache MP cache server: "
+                "runs LOOKUP / STORE / RETRIEVE against a live MP server "
+                "and verifies KV cache checksums."
+            ),
+        )
+        TestCacheCommand().add_arguments(parser)
+        parser.set_defaults(func=self.execute)
+
+    def _bench_kvcache(self, args: argparse.Namespace) -> None:
+        """Dispatch ``lmcache bench kvcache`` to ``TestCacheCommand``."""
+        TestCacheCommand().execute(args)
+
     def execute(self, args: argparse.Namespace) -> None:
-        handlers = {"engine": self._bench_engine}
+        handlers = {
+            "engine": self._bench_engine,
+            "kvcache": self._bench_kvcache,
+        }
         handler = handlers.get(args.bench_target)
         if handler is None:
             print(
